Use event's PhysicalResourceId if provided
Fixes issue #2.

If no user-provided physical resource ID is found, try to use the CloudFormation-provided physical resource ID, falling back to the mostly random context's log stream name only as last resort. 

This will cause the CloudFormation replace-on-update logic to behave as expected and not "replace" custom resources by deleting them in the update cleanup stage, unless the user really intends it to be so.